+++ src/main/java/com/Appointment/AppointmentService.java
@@ -161,7 +161,7 @@ public class AppointmentService {
         transactionDao.delete(transaction);
     }
 
-    public String rateAppointment(int appointment_id, int score) {
+    public String rateAppointment(int appointment_id, double score) {
         Appointment appointment = appointmentDao.findById(appointment_id);
         appointment.setScore(score);
         appointment.setQualified(true);
